backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.3
pytz==2024.2
pyodbc==5.1.0
psutil==6.1.0
email-validator==2.2.0
fastapi-limiter==0.1.6
schedule==1.2.2
# PostgreSQL dependencies
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
tenacity==8.2.3
